Create Automated Guided Vehicle with Temperature Sensor
This Project is to build an obstacle sensing and temperature detecting Robot. The main objective of the project is to detect the temperature using robot which is automatically controlled using obstacle sensors. The robot can move through any path whose motion is controlled automatically using the obstacle sensors interfaced by a software program. When the robot detects a change in the temperature using the temperature sensor, it compares the detected temperature with a fixed value that is set by default in the software program. If the temperature exceeds the fixed temperature, the cooling fan is switched on and if the temperature reduces below the default temperature, the cooling fan is switched off. The motion of the robot is controlled by the software program depending on which obstacle sensors receive/transmit data first.